+++ apps/modernjs-ssr/dynamic-nested-remote/@mf-types/dynamic_remote/compiled-types/Image.d.ts
@@ -1,4 +1,0 @@
-declare const _default: ({ text }: {
-    text: string;
-}) => JSX.Element;
-export default _default;