+++ packages/enhanced/src/wrapper/HoistContainerReferencesPlugin.ts
@@ -7,10 +7,21 @@ export default class HoistContainerReferencesPlugin
   implements WebpackPluginInstance
 {
   name: string;
-  private containerName?: string | undefined;
+  private containerName: string;
+  private entryFilePath?: string;
+  private bundlerRuntimeDep?: string;
+  private explanation: string;
 
-  constructor(name?: string | undefined) {
-    this.containerName = name;
+  constructor(
+    name?: string,
+    entryFilePath?: string,
+    bundlerRuntimeDep?: string,
+  ) {
+    this.containerName = name || 'no known chunk name';
+    this.entryFilePath = entryFilePath;
+    this.bundlerRuntimeDep = bundlerRuntimeDep;
+    this.explanation =
+      'Bundler runtime path module is required for proper functioning';
     this.name = PLUGIN_NAME;
   }
 
@@ -20,6 +31,10 @@ export default class HoistContainerReferencesPlugin
     const CoreHoistContainerReferencesPlugin =
       require('../lib/container/HoistContainerReferencesPlugin')
         .default as typeof import('../lib/container/HoistContainerReferencesPlugin').default;
-    new CoreHoistContainerReferencesPlugin().apply(compiler);
+    new CoreHoistContainerReferencesPlugin(
+      this.containerName,
+      this.entryFilePath,
+      this.bundlerRuntimeDep,
+    ).apply(compiler);
   }
 }