+++ packages/enhanced/test/ConfigTestCases.embedruntime.js
@@ -1,0 +1,8 @@
+const { describeCases } = require('./ConfigTestCases.template');
+
+describeCases({
+  name: 'ConfigTestCases',
+  federation: {
+    federationRuntime: 'hoisted',
+  },
+});