+++ packages/modernjs/src/cli/configPlugin.ts
@@ -54,6 +54,7 @@ export const moduleFederationConfigPlugin = (
   userConfig: InternalModernPluginOptions,
 ): CliPlugin<AppTools> => ({
   name: '@modern-js/plugin-module-federation-config',
+  pre: ['@modern-js/plugin-initialize'],
   post: ['@modern-js/plugin-module-federation'],
   setup: async ({ useConfigContext, useAppContext }) => {
     const modernjsConfig = useConfigContext();
@@ -127,6 +128,10 @@ export const moduleFederationConfigPlugin = (
               FEDERATION_IPV4: JSON.stringify(ipv4),
               REMOTE_IP_STRATEGY: JSON.stringify(userConfig.remoteIpStrategy),
             },
+            enableAsyncEntry:
+              bundlerType === 'rspack'
+                ? modernjsConfig.source?.enableAsyncEntry ?? true
+                : modernjsConfig.source?.enableAsyncEntry,
           },
           dev: {
             assetPrefix: modernjsConfig?.dev?.assetPrefix