+++ packages/runtime/src/utils/env.ts
@@ -1,3 +1,5 @@
+export { isBrowserEnv, isDebugMode } from '@module-federation/sdk';
+
 export function isDevelopmentMode(): boolean {
   return true;
 }