+++ packages/sdk/rollup.config.js
@@ -6,12 +6,16 @@ module.exports = (rollupConfig, _projectOptions) => {
       targets: [{ src: 'packages/sdk/LICENSE', dest: 'packages/sdk/dist' }],
     }),
   );
-  // rollupConfig.plugins.push({
-  //   name: 'custom-dynamic-import',
-  //   renderDynamicImport({ moduleId }) {
-  //     return { left: 'import(', right: ')' };
-  //   },
-  // });
+
+  rollupConfig.external = [/@module-federation/];
+  rollupConfig.output = {
+    ...rollupConfig.output,
+    manualChunks: (id) => {
+      if (id.includes('@swc/helpers')) {
+        return 'polyfills';
+      }
+    },
+  };
 
   return rollupConfig;
 };