+++ apps/runtime-demo/3008-runtime-remote/CHANGELOG.md
@@ -1,5 +1,31 @@
 # 3008-runtime-remote
 
+## 1.0.34
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.4
+
+## 1.0.33
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.3
+
+## 1.0.32
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.2
+
+## 1.0.31
+
+### Patch Changes
+
+- Updated dependencies [2855583]
+- Updated dependencies [2855583]
+  - @module-federation/enhanced@0.6.1
+
 ## 1.0.30
 
 ### Patch Changes