+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.6.4
+
+## 0.6.3
+
+## 0.6.2
+
+## 0.6.1
+
 ## 0.6.0
 
 ## 0.5.2