+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,31 @@
 # [3.1.0](https://github.com/module-federation/core/compare/utils-3.0.2...utils-3.1.0) (2023-10-26)
 
+## 3.1.10
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.4
+
+## 3.1.9
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.3
+
+## 3.1.8
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.2
+
+## 3.1.7
+
+### Patch Changes
+
+- Updated dependencies [2855583]
+- Updated dependencies [813680f]
+  - @module-federation/sdk@0.6.1
+
 ## 3.1.6
 
 ### Patch Changes